portfolio-ai/backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
langchain-huggingface
sentence-transformers
faiss-cpu
pydantic
requests
tiktoken
psutil
numpy<2.0.0

